requirements.txt
torch>=2.0.0
torch-geometric>=2.5.0
pyg-lib>=0.4.0
torch-sparse>=0.6.17
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.23.0
networkx>=3.1
requests>=2.28.0
neo4j>=5.12.0
python-dotenv>=1.0.0
gradio>=4.0.0
langchain>=0.1.17
langchain-community>=0.0.32
langchain-huggingface>=0.0.1
langchain-ollama>=0.0.6
langchain-text-splitters>=0.0.1
langchain-neo4j>=0.0.2
langchain-experimental>=0.0.47
